+++ app/services/admin/contacts/update_info.rb
@@ -1,6 +1,7 @@
 class Admin::Contacts::UpdateInfo < Admin::Contacts
   def do
-    if Contact::SetZazoInfoByContact.new(@contact).do
+    if Contact::Update::FindZazoContact.new(@contact).do &&
+       Contact::Update::UpdateZazoInfo.new(@contact).do
       [true, 'Contact info was updated']
     else
       [false, 'Something was wrong']