+++ app/workers/resque_worker/update_mkey_defined_contact.rb
@@ -2,7 +2,7 @@ class ResqueWorker::UpdateMkeyDefinedContact
   @queue = :update_contacts
 
   def self.perform(contact_id)
-    WriteLog.info(self, "resque job was executed for contact_id=#{contact_id}")
+    WriteLog.info(self, "started; contact_id: '#{contact_id}'")
     contact = Contact.find(contact_id)
     Contact::UpdateZazoInfo.new(contact).do
     Score::CalculationByContact.new(contact).do