+++ app/workers/resque_worker/update_owner_contacts_only.rb
@@ -2,7 +2,8 @@ class ResqueWorker::UpdateOwnerContactsOnly
   @queue = :update_contacts
 
   def self.perform(owner_mkey)
-    WriteLog.info(self, "resque job was executed for owner_mkey=#{owner_mkey}")
-    Contact::SetZazoInfoByOwner.new(owner_mkey).do
+    WriteLog.info(self, "started; owner: '#{owner_mkey}'")
+    owner = Owner.new(owner_mkey)
+    owner.contacts_actions.find_contact_and_update_info
   end
 end