+++ spec/services/web_client/add_contact_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe WebClient::AddContact do
-
-end